refactor: folder distribution
