+++ src/backend/go.mod
@@ -1,0 +1,8 @@
+module PC4
+
+go 1.23.0
+
+require (
+	github.com/gorilla/mux v1.8.1 // indirect
+	github.com/gorilla/websocket v1.5.3 // indirect
+)
