--- src/dataset/ratings2.csv
@@ -1,18 +1,0 @@
-userId,movieId,rating,timestamp
-1,110,5.0,1425941529
-1,148,5.0,1425941529
-1,150,5.0,1425941529
-2,110,5,1425942435
-2,148,5,1425942435
-2,150,5,1425942435
-2,160,5,1425942435
-2,170,5,1425942435
-3,110,5.0,1425941523
-3,180,5.0,1425941523
-3,150,5.0,1425941523
-3,200,5.0,1425941523
-4,1221,1,1425941546
-5,1246,1,1425941556
-6,1968,1,1425942148
-7,2762,1,1425941300
-8,2918,1,1425941593
